Create Reverse a string by word
The code takes a sentence "This is javascript code" and splits it into an array of words using the split() method.
 It then declares an empty string reverseWord and iterates through the arrSentence array in reverse order using a for loop starting from the last index ( arrSentence.length - 1) and ending at the first index ( 0).
Inside the loop, it appends each word of arrSentence into the reverseWord string along with an additional space character. 
The resulting string will have an extra space at the end.
Finally, it logs the reverseWord string which contains the words of the sentence in reverse order separated by two spaces.