Create Reverse a string
This code takes a string "This is javascript code" and converts it into an array using the Array.from() method. 
It then initializes an empty string called reverseString. 
A for loop is used to iterate over the array from the end to the beginning using the arrSentence.length-1 to get the last index of the array. 
Inside the loop, it concatenates each character to the reverseString string starting from the end.
 The loop runs until the index becomes less than 0.
Finally, the reverseString string containing the reversed string is printed using the console.log() function.