Create Remove duplicate elements in Array
Code creates an array called arr with a mixture of numbers and strings. 
Then, it applies the filter() method on this array to create a new array called uniArray, 
which contains only the unique elements of the original array. 
The filter method uses the indexOf() method to check if the current element's index is the same as the first occurrence of the element in the array. 
If they are the same, it means that the element is unique, so it is added to the uniArray. 
Finally, the uniArray is printed to the console .