Create Find Minimum number in the array Without inbuilt method
Code finds the minimum value in an array of numbers arr using a for loop and a conditional statement. 
The variable min is initialized to the first element of the array arr[0]. 
The loop then iterates through the remaining elements of the array using i as the index. 
For each iteration, the current element arr[i] is compared with the current minimum value min using a conditional statement.
 If the current element is less than the current minimum, then the current element becomes the new minimum value. 
Finally, the code outputs the minimum value using console.log().