Create Find Maximum number in the array Without using inbuilt method
Code finds the maximum value in an array of numbers arr. 
It initializes a variable max with the first element of the array.
 Then it iterates over the array using a for loop and compares each element to the current max. 
If an element is greater than the current max, the max variable is updated to hold that element. 
Finally, the value of max is printed to the console using console.log(max).