Create Find Minimum number in the array using inbuilt method
Code uses the Math.min() method to find the smallest number in an array arr. 
The spread operator ... is used to spread the array arr into individual elements so that they can be passed as separate arguments to the Math.min() method. 
The returned value is assigned to the variable min.
 which is then printed to the console using console.log().